fix(license): Fixed the NullPointerException and addressed code duplication
