+++ frontend/sw360-portlet/src/main/java/org/eclipse/sw360/portal/portlets/admin/LicenseAdminPortlet.java
@@ -148,22 +148,4 @@ public class LicenseAdminPortlet extends Sw360Portlet {
         final UploadPortletRequest uploadPortletRequest = PortalUtil.getUploadPortletRequest(request);
         return uploadPortletRequest.getFileAsStream(fileUploadFormId);
     }
-
-    public void backUpLicenses(ResourceRequest request, ResourceResponse response) throws IOException, TException {
-        final LicenseService.Iface licenseClient = thriftClients.makeLicenseClient();
-        Map<String, InputStream> fileNameToStreams = (new LicsExporter(licenseClient)).getFilenameToCSVStreams();
-
-        final ByteArrayOutputStream outB = new ByteArrayOutputStream();
-        final ZipOutputStream zipOutputStream = new ZipOutputStream(outB);
-
-        for (Map.Entry<String, InputStream> entry : fileNameToStreams.entrySet()) {
-            ZipTools.addToZip(zipOutputStream, entry.getKey(), entry.getValue());
-        }
-
-        zipOutputStream.flush();
-        zipOutputStream.close(); // this closes outB
-
-        final ByteArrayInputStream zipFile = new ByteArrayInputStream(outB.toByteArray());
-        PortletResponseUtil.sendFile(request, response, "LicensesBackup.lics", zipFile, "application/zip");
-    }
 }
